Use tri-state channel status
Channels can now be soft- or hard-disabled or enabled. Hard-disabling
is a manual process and a hard-disabled channel can only be re-enabled
manually. Soft-disabling can be done automatically by balance if it
finds a channel to be degraded. Soft-disabled channels can be re-
enabled by balance if the channel becomes functional again. But this
feature is yet to be implemented.