Implement monitoring support
This change implements channel monitoring. In order to try and keep
as much of the original code as possible, monitoring is implemented
by forking off a separate monitoring process that communicates with
the master process using the same means (shared memory) as the shell
and connection processes.

There is support for both simple TCP connect checking as well as
arbitrary health check commands that can be run by the monitoring
process.