Ensure streaming processes can get killed (in a safe way)
I noticed during testing that after my changes to catch TERM/QUIT/INT
it was no longer possible to kill the streaming kids of the master
process. This was obviously bad. However, just removing the signal
handlers for the streaming process is bad as well as killing such a
process would cause the connection counters to become invalid.

So this change handles signals in the streaming process in much the
same way as they're handled in the master and monitoring process and
attempts to shut the connection down gracefully when catching a signal.

This is still not a perfect solution as inconsistencies can arise when
the process is killed with a signal that cannot be caught (SIGKILL).